+++ client/wallet/connector.go
@@ -9,8 +9,10 @@ import (
 // Connector represents an interface that defines how the wallet interacts with the network. A wallet can either be used
 // locally on a server or it can connect remotely using the web API.
 type Connector interface {
-	UnspentOutputs(addresses ...address.Address) (unspentOutputs map[address.Address]map[ledgerstate.OutputID]*Output, err error)
+	UnspentOutputs(addresses ...address.Address) (unspentOutputs OutputsByAddressAndOutputID, err error)
 	SendTransaction(transaction *ledgerstate.Transaction) (err error)
-	RequestFaucetFunds(address address.Address) (err error)
+	RequestFaucetFunds(address address.Address, powTarget int) (err error)
 	GetAllowedPledgeIDs() (pledgeIDMap map[mana.Type][]string, err error)
+	GetTransactionInclusionState(txID ledgerstate.TransactionID) (inc ledgerstate.InclusionState, err error)
+	GetUnspentAliasOutput(address *ledgerstate.AliasAddress) (output *ledgerstate.AliasOutput, err error)
 }