+++ packages/remotelogmetrics/events.go
@@ -47,6 +47,44 @@ type FPCConflictRecord struct {
 	Opinions []int32 `json:"opinions" bson:"opinions"`
 	// Outcome defines final opinion of the conflict.
 	Outcome int32 `json:"outcome,omitempty" bson:"outcome,omitempty"`
-	// Time defines the time when the conflict has been finalized.
+	// Time defines the time when the voting round has been finalized.
 	Time time.Time `json:"datetime" bson:"datetime"`
+	// ConflictCreationTime points to time when the context has been created
+	ConflictCreationTime time.Time `json:"conflictStart" bson:"conflictStart"`
+}
+
+// TransactionMetrics defines the transaction metrics record to sent be to remote logger.
+type TransactionMetrics struct {
+	Type               string    `json:"type" bson:"type"`
+	MessageID          string    `json:"messageID" bson:"messageID"`
+	TransactionID      string    `json:"transactionID" bson:"transactionID"`
+	IssuedTimestamp    time.Time `json:"issuedTimestamp" bson:"issuedTimestamp"`
+	SolidTimestamp     time.Time `json:"solidTimestamp" bson:"solidTimestamp"`
+	ScheduledTimestamp time.Time `json:"scheduledTimestamp" bson:"scheduledTimestamp"`
+	BookedTimestamp    time.Time `json:"bookedTimestamp" bson:"bookedTimestamp"`
+	ConfirmedTimestamp time.Time `json:"confirmedTimestamp" bson:"confirmedTimestamp"`
+}
+
+// DRNGMetrics defines the DRNG metrics record to sent be to remote logger.
+type DRNGMetrics struct {
+	Type              string    `json:"type" bson:"type"`
+	InstanceID        uint32    `json:"instanceID" bson:"instanceID"`
+	Round             uint64    `json:"round" bson:"round"`
+	IssuedTimestamp   time.Time `json:"issuedTimestamp" bson:"issuedTimestamp"`
+	ReceivedTimestamp time.Time `json:"receivedTimestamp" bson:"receivedTimestamp"`
+}
+
+// StatementLog defines the statement metrics record to sent be to remote logger.
+type StatementLog struct {
+	NodeID       string    `json:"nodeID"`
+	MsgID        string    `json:"msgID"`
+	IssuerID     string    `json:"issuerID"`
+	IssuedTime   time.Time `json:"issuedTime"`
+	ArrivalTime  time.Time `json:"arrivalTime"`
+	SolidTime    time.Time `json:"solidTime"`
+	DeltaArrival int64     `json:"deltaArrival"`
+	DeltaSolid   int64     `json:"deltaSolid"`
+	Clock        bool      `json:"clock"`
+	Sync         bool      `json:"sync"`
+	Type         string    `json:"type"`
 }