+++ plugins/dashboard/frontend/src/app/stores/ExplorerStore.tsx
@@ -7,7 +7,8 @@ import {
     DrngSubtype,
     PayloadType,
     TransactionPayload,
-    getPayloadType
+    getPayloadType,
+    Output
 } from "app/misc/Payload";
 import * as React from "react";
 import {Link} from 'react-router-dom';
@@ -44,21 +45,18 @@ export class Message {
     futureMarkers: string;
 }
 
-class AddressResult {
+export class AddressResult {
     address: string;
-    output_ids: Array<Output>;
+    explorerOutputs: Array<ExplorerOutput>;
 }
 
-class Output {
-    id: string;
-    transaction_id: string;
-    type: string;
-    index: number;
-    balances: Array<Balance>;
-    inclusion_state: InclusionState;
-    consumer_count: number;
-    solidification_time: number;
-    pending_mana: number;
+export class ExplorerOutput {
+    id: OutputID;
+    output: Output;
+    metadata: OutputMetadata
+    inclusionState: InclusionState;
+    txTimestamp: number;
+    pendingMana: number;
 }
 
 class OutputID {
@@ -67,7 +65,7 @@ class OutputID {
     outputIndex: number;
 }
 
-class OutputMetadata {
+export class OutputMetadata {
     outputID: OutputID;
     branchID: string;
     solid: boolean;
@@ -125,12 +123,7 @@ class BranchConflicts {
     conflicts: Array<BranchConflict>
 }
 
-class Balance {
-    value: number;
-    color: string;
-}
-
-class InclusionState {
+export class InclusionState {
 	liked: boolean;
 	rejected: boolean;
 	finalized: boolean;
@@ -272,11 +265,11 @@ export class ExplorerStore {
             }
             let tx = await res.json()
             for(let i = 0; i < tx.inputs.length; i++) {
-                let inputID = tx.inputs[i].referencedOutputID ? tx.inputs[i].referencedOutputID.base58 : GenesisMessageID
+                let inputID = tx.inputs[i] ? tx.inputs[i].referencedOutputID.base58 : GenesisMessageID
                 try{
                     let referencedOutputRes = await fetch(`/api/output/${inputID}`)
                     if (referencedOutputRes.status === 200){
-                        tx.inputs[i].referencedOutput = await referencedOutputRes.json()
+                        tx.inputs[i].output = await referencedOutputRes.json()
                     }
                 }catch(err){
                     // ignore
@@ -561,7 +554,7 @@ export class ExplorerStore {
                 <tr key={msg.id}>
                     <td>
                         <Link to={`/explorer/message/${msg.id}`}>
-                            {msg.id.substr(0, 35)}
+                            {msg.id}
                         </Link>
                     </td>
                     <td>