+++ plugins/remotelogmetrics/plugin.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/remotelogmetrics"
 	"github.com/iotaledger/goshimmer/packages/shutdown"
+	"github.com/iotaledger/goshimmer/plugins/drng"
 	"github.com/iotaledger/goshimmer/plugins/messagelayer"
 	"github.com/iotaledger/goshimmer/plugins/remotelog"
 )
@@ -39,6 +40,9 @@ func Plugin() *node.Plugin {
 func configure(_ *node.Plugin) {
 	configureSyncMetrics()
 	configureFPCConflictsMetrics()
+	configureDRNGMetrics()
+	configureTransactionMetrics()
+	configureStatementMetrics()
 }
 
 func run(_ *node.Plugin) {
@@ -75,4 +79,16 @@ func configureFPCConflictsMetrics() {
 	metricsLogger := newFPCMetricsLogger()
 	messagelayer.Voter().Events().Finalized.Attach(events.NewClosure(metricsLogger.onVoteFinalized))
 	messagelayer.Voter().Events().RoundExecuted.Attach(events.NewClosure(metricsLogger.onVoteRoundExecuted))
+}
+
+func configureDRNGMetrics() {
+	drng.Instance().Events.Randomness.Attach(events.NewClosure(onRandomnessReceived))
+}
+
+func configureTransactionMetrics() {
+	messagelayer.Tangle().LedgerState.UTXODAG.Events.TransactionConfirmed.Attach(events.NewClosure(onTransactionConfirmed))
+}
+
+func configureStatementMetrics() {
+	messagelayer.Tangle().ConsensusManager.Events.StatementProcessed.Attach(events.NewClosure(onStatementReceived))
 }